Make generic Post-Form to use it in Add and Edit Post Pages
