+++ src/app/components/post-owner/post-form/post-form.component.css
@@ -1,0 +1,3 @@
+mat-form-field {
+    width: 90%;
+}
